requirements.txt
flake8
pytest
fastapi
uvicorn
numpy
pandas
requests
pydantic
paho-mqtt


